+++ src/screens/Products/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ProductsCarousel } from "@/features/ProductsCarousel";
 import { SelectProduction } from "@/features/SelectProduction";
 import { ProductDetailItem } from "@/ui/ProductDetailItem";
-import { Production, Props as ProductionProps } from "@/ui/Production";
+import { ProductionDetailProps as ProductionProps } from "@/ui/Production";
 
 type Props = {
   products: ProductionProps[];
@@ -12,7 +12,7 @@ export const ProductsScreen: React.FC<Props> = ({ products }) => {
   return (
     <div>
       <h1>Products</h1>
-      <ProductDetailItem products={products} />
+      <ProductsCarousel products={products} />
       <SelectProduction products={products} />
     </div>
   );
